move common result page functions to own file
